Add docker-compose file
