+++ .env example
@@ -1,0 +1,4 @@
+IP=127.0.0.1
+MONGO_PORT=
+COLLECTION_NAME=
+JWT_SECRET=
