+++ .gitignore
@@ -1,3 +1,3 @@
 node_modules
-config.js
+.env
 dist
