+++ deploy/mongo/Dockerfile
@@ -1,0 +1,1 @@
+FROM mongo:4.2.0
