+++ deploy/mongo/README.md
@@ -1,0 +1,19 @@
+# Деплой проекта
+
+## Подготовительный этап
+
+### Подключение MongoDB локально из контейнера с помощью docker-compose
+
+**Для чего это надо?**
+
+Во-первых это облегчит последующий деплой на vps сервер
+Во-вторых это помогает разворачивать MongoDB локально с помощью одной команды
+### Шаги для выполнения
+1.  Скачать [docker-compose](https://docs.docker.com/compose/install/) локально(для своей ОС)
+2. Создать локально `.env`  файл(скопировать содержимое `.env.example`) в котором указать переменные для запуска локальной MongoDB.
+3. Выполнить в корне проекта команду
+```bash
+docker-compose up -d
+```
+4. Протестировать подключение к MongoDB через MongoDB Compas, передав теже самые параметры для подключение,  которые раннее были указаны в  файле `.env`
+5. Дальше можно тестировать уже само API проекта
