+++ package.json
@@ -16,6 +16,7 @@
     "compression": "^1.7.4",
     "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
+    "dotenv": "^16.0.1",
     "express": "^4.18.1",
     "express-jwt": "^7.7.5",
     "helmet": "^5.1.0",
